backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
pdfplumber
google-genai
SQLAlchemy
aiosqlite
